Reject space before filename with folders
If the layer filename is of format: [folder]/[whitespace][filename]…
parser throws an error, which gets logged to errors.txt.

Also update the unit tests so they call a unit test specific function
which catches the error thrown by parser so unit tests can check for
error messages. This needs to be refactored later to apply to all
parsing rules.